ai-service/requirements.txt
fastapi
uvicorn
pymupdf
faiss-cpu
numpy
requests
langchain-text-splitters
supabase
python-dotenv
scikit-learn
openai